Fix a bug with testing the flexbuffers serializer
The flexbuffers serializer may lose empty structures at the end of an AoS.

For example: serializing a core profiles IDS with 5 profiles_1d structures:
    [empty, filled, empty, filled, empty]
will discard the empty aos structure at the end. The empty ones in the center are not affected.

This is an edge case which doesn't make much sense. However, when filling structures with random data, we sometimes end up in this situation.
This commit updates `fill_with_random_data` to delete empty structures at the back of an AoS.